Helios-AI-2026-main/backend/requirements.txt
firebase-admin
google-genai
python-dotenv
requests
